Move BolusWindow sub-composables into component files
